mysql升级到8.0.11版本，实现spring Aspect功能

+++ src/main/java/com/hu/ssm/base/mapper/SysUserMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface SysUserMapper {
-    int countByExample(SysUserExample example);
+    long countByExample(SysUserExample example);
 
     int deleteByExample(SysUserExample example);
